feat(chatroom): basic socket script finish, need revise and test
