+++ src/backend-ts/package.json
@@ -26,7 +26,6 @@
     "mysql2": "^3.11.4",
     "node-cron": "^3.0.3",
     "redis": "^4.7.0",
-    "socket.io": "^4.8.1",
     "systeminformation": "^5.23.5"
   },
   "devDependencies": {
