+++ src/backend-ts/utils/redis.ts
@@ -1,7 +1,5 @@
 import Redis from "ioredis";
 
-const redis = new Redis(6380);
-
-export const REDIS_URL = "http://localhost:6380";
+const redis = new Redis("redis");
 
 export default redis;
