+++ src/chat/dockerfile
@@ -1,0 +1,13 @@
+FROM denoland/deno:2.0.0
+
+USER deno
+
+WORKDIR /app
+
+COPY . /app/
+
+RUN deno install
+
+EXPOSE 8080
+
+CMD ["deno", "run", "--allow-all", "--unstable-sloppy-imports", "src/index.ts"]
